+++ app/models/ta_membership.rb
@@ -22,14 +22,15 @@
 #
 # rubocop:enable Layout/LineLength, Lint/RedundantCopDisableDirective
 class TaMembership < Membership
-  validate :must_be_a_ta
+  validate :must_be_course_staff
 
   after_create { Repository.get_class.update_permissions }
   after_destroy { Repository.get_class.update_permissions }
 
-  def must_be_a_ta
-    if role && !role.is_a?(Ta)
-      errors.add(:base, :not_ta)
+  def must_be_course_staff
+    # Use instance_of? to exclude AdminRole, which is a subclass of Instructor.
+    if role && !role.is_a?(Ta) && !role.instance_of?(Instructor)
+      errors.add(:base, :not_course_staff)
       false
     end
   end
@@ -44,15 +45,15 @@ class TaMembership < Membership
     end
     new_ta_memberships = []
     groupings = assignment.groupings.joins(:group).pluck('groups.group_name', :id).to_h
-    graders = assignment.course.tas.joins(:user).pluck('users.user_name', :id).to_h
+    course_staff = assignment.course.course_staff.joins(:user).pluck('users.user_name', :id).to_h
     result = MarkusCsv.parse(csv_data) do |row|
       raise CsvInvalidLineError if row.empty?
       raise CsvInvalidLineError if groupings[row[0]].nil?
 
       row.drop(1).each do |grader_name|
-        unless graders[grader_name].nil?
+        unless course_staff[grader_name].nil?
           new_ta_memberships << {
-            role_id: graders[grader_name],
+            role_id: course_staff[grader_name],
             grouping_id: groupings[row[0]],
             type: 'TaMembership'
           }